+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `QuestionAndAnswer` node for parsing Q&A lines in minimarkdown
+- Typing support for question-and-answer format
+- `QuestionAndAnswerTemplate` for creating Q&A challenges
+
 ## [0.10.0] - 2019-09-14
 
 ### Changed